requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
ffmpeg==1.4
ffmpeg-downloader==0.3.0
frozenlist==1.4.1
idna==3.6
magic-filter==1.0.12
multidict==6.0.5
openai==0.27.0
packaging==23.2
platformdirs==4.2.0
pydantic==2.5.3
pydantic_core==2.14.6
pydub==0.25.1
python-dotenv==1.0.1
requests==2.31.0
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.2.0
yarl==1.9.4
